fix: update Sunshine and NVIDIA driver versions; enhance Steam package installation process

+++ overlay/usr/bin/install_steam_pkg.sh
@@ -2,8 +2,6 @@
 
 echo "**** Installing/upgrading Steam package ****"
 
-mkdir -p ${USER_HOME}/.steam/debian-installation/package
-
 pkg=(
 tenfoot_images_all.zip.vz.193cb8c4eb4446698ea2c0a9e8c4e6b6a623dac7_5572671
 resources_all.zip.vz.3d492fce87e5ccddbb855f26680b0c6798901010_2867227
@@ -23,22 +21,58 @@ bins_misc_ubuntu12.zip.vz.72be98aad8e15e7b92bfccfe7c4268b720870b9c_18827909
 webkit_ubuntu12.zip.vz.57c0e0d3866ff0cdaa4af1f0fb2a0393a31a7906_79943183
 runtime_scout_ubuntu12.zip.2422dc5093c67022c86b17493e49f66124f182d0
 runtime_sniper_ubuntu12.zip.328e060d569aa12d70746dab1f74cd54196edf9c
+miles_ubuntu12.zip.vz.5093ef941e6e5195a60ab3259077694dec994016_295496
 )
+
+if [[ ${{PREINSTALL_STEAM_PACKAGE:-}} == "true" ]]; then
+    echo "Pre-installing Steam package ... "
+    for p in ${pkg[@]}; do
+        if [[ ! -f /usr/local/src/$p ]]; then
+            echo "Downloading [${p}] ... "
+            checksum=$(echo $p|awk -F'.' '{print $NF}'|awk -F'_' '{print $1}')
+            retry=10
+            while [[ retry -gt 0 ]]; do
+                wget -O /usr/local/src/$p http://media.steampowered.com/client/$p
+                sum=$(sha1sum /usr/local/src/$p|awk '{print $1}')
+                if [[ x"$sum" == x"$checksum" ]]; then
+                    break
+                fi
+
+                ((retry--))
+            done
+
+            if [[ ! -f /usr/local/src/$p ]]; then
+                echo "Failed to download [${p}] after multiple attempts, exiting."
+                exit 1
+            fi
+        fi
+    done
+
+    echo "Pre-install Steam package DONE"
+    exit 0
+fi
+
+mkdir -p ${USER_HOME}/.steam/debian-installation/package
 
 for p in ${pkg[@]}; do
     if [[ ! -f ${USER_HOME}/.steam/debian-installation/package/$p ]]; then
-        echo "Downloading [${p}] ... "
-        checksum=$(echo $p|awk -F'.' '{print $NF}'|awk -F'_' '{print $1}')
-        retry=10
-        while [[ retry -gt 0 ]]; do
-            wget -O ${USER_HOME}/.steam/debian-installation/package/$p http://media.steampowered.com/client/$p
-            sum=$(sha1sum ${USER_HOME}/.steam/debian-installation/package/$p|awk '{print $1}')
-            if [[ x"$sum" == x"$checksum" ]]; then
-                break
-            fi
+        if [[ -f /usr/local/src/$p ]]; then
+            echo "Copying pre-downloaded package [${p}] ... "
+            cp -v /usr/local/src/$p ${USER_HOME}/.steam/debian-installation/package/$p
+        else
+            echo "Downloading [${p}] ... "
+            checksum=$(echo $p|awk -F'.' '{print $NF}'|awk -F'_' '{print $1}')
+            retry=10
+            while [[ retry -gt 0 ]]; do
+                wget -O ${USER_HOME}/.steam/debian-installation/package/$p http://media.steampowered.com/client/$p
+                sum=$(sha1sum ${USER_HOME}/.steam/debian-installation/package/$p|awk '{print $1}')
+                if [[ x"$sum" == x"$checksum" ]]; then
+                    break
+                fi
 
-            ((retry--))
-        done
+                ((retry--))
+            done
+        fi
     fi
 done
 
@@ -47,6 +81,6 @@ if [[ ! -f ${USER_HOME}/.steam/debian-installation/compatibilitytools.d/GE-Proto
     mkdir -p ${USER_HOME}/.steam/debian-installation/compatibilitytools.d
     cd ${USER_HOME}/.steam/debian-installation/compatibilitytools.d
     tar zxvf /usr/local/src/GE-Proton9-15.tar.gz
-if
+fi
 
 echo "DONE"
